# docs/api/sancho-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sancho.API | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5099/"
    }
  ],
  "paths": {
...
    "/api/user/me": {
      "get": {
        "tags": [
          "UserEndpoints"
        ],
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "UserEndpoints"
        ],
        "operationId": "UpdateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/user/me/avatar/upload-url": {
      "post": {
        "tags": [
          "UserEndpoints"
        ],
        "operationId": "GetAvatarUploadUrl",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUploadUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/user/me/avatar/confirm": {
      "post": {
        "tags": [
          "UserEndpoints"
        ],
        "operationId": "ConfirmAvatar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAvatarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/user/me/permissions": {
      "get": {
        "tags": [
          "UserEndpoints"
        ],
        "operationId": "GetUserPermissions",
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/identity/users/{userId}": {
      "get": {
        "tags": [
          "IdentityEndpoints"
        ],
        "operationId": "GetIdentityUserDetail",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvatarUploadUrlRequest": {
        "required": [
          "contentType"
        ],
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string"
          }
        }
      },
      "ConfirmAvatarRequest": {
        "required": [
          "filePath"
        ],
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string"
          }
        }
      },
...
      "UpdateProfileRequest": {
        "required": [
          "displayName",
          "bio",
          "locale"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}